fix: remove all non-ASCII characters to bypass parser crash

+++ 211-command-usb/device/device.c
@@ -39,26 +39,26 @@ void dev_info(void)
     
     printf("%-15s 0x%08x %5u\n", 
            "device_card", 
-           &device_card, 
+           (uint32_t)&device_card, 
            sizeof(device_card));
 
     printf("- %-13s 0x%08x %5u %6u 0x%08x\n",
            "version",
-           &device_card.version,
+           (uint32_t)&device_card.version,
            sizeof(device_card.version),
            offsetof(struct info_t, version),
            device_card.version);
 
     printf("- %-13s 0x%08x %5u %6u %s\n",
            "name",
-           device_card.name,
+           (uint32_t)device_card.name,
            sizeof(device_card.name),
            offsetof(struct info_t, name),
            device_card.name);
 
     printf("- %-13s 0x%08x %5u %6u %u\n",
            "revision",
-           &device_card.revision,
+           (uint32_t)&device_card.revision,
            sizeof(device_card.revision),
            offsetof(struct info_t, revision),
            device_card.revision);
